Freshness is asked of the module that owns Cache-Control
`helpers/mod.rs` states the shelving rule: every module is named for a
question, none for a document. `headers.rs` is named for a data structure, and
had accumulated whatever arrives in a `HeaderMap` — eleven unrelated questions
across 2344 lines.

The four freshness functions are the clearest of them, because each was already
calling this module and adding a policy on top: three opened with
`cache_control::forbids_storage_or_reuse` and those calls are now local, and
`compute_freshness_lifetime` carried § 4.2.1's "use the first match" ordering,
which is a statement about directives.

`estimated_age` reads `Age`, not `Cache-Control`, and comes anyway. The question
is freshness and `Age` is one of its two inputs; shelving it by the field it
happens to read is what put it in a module named after a header map.

Five rules updated to the new path. The 4 citations moved with their functions —
`headers.rs` 56 → 52, `cache_control.rs` 10 → 14, total unchanged at 2383, with
103 entries renumbered because the generated file records line numbers.

headers.rs: 2344 → 2224 non-test lines.
