+++ lint-http-rules/src/helpers/cache_control.rs
@@ -214,6 +214,139 @@ pub fn forbids_storage_or_reuse(headers: &HeaderMap) -> bool {
     directives(headers).any(|d| d.is_unqualified("no-store") || d.is_unqualified("no-cache"))
 }
 
+// ── Freshness, computed from the directives above ────────────────────
+//
+// These four came from `helpers/headers.rs`, which is named for a data
+// structure rather than for a question and had collected every field that
+// arrives in a `HeaderMap`. Each one was already asking this module its
+// question and adding a caching policy on top — three of the four opened by
+// calling `forbids_storage_or_reuse`, and those calls are now local.
+//
+// `estimated_age` reads `Age` rather than `Cache-Control`, and belongs here on
+// the same rule the module doc states: the question is freshness, and `Age` is
+// one of its two inputs. Shelving it by the field it happens to read would put
+// it back in a module named after a header map.
+
+/// Extract the numeric value of a `max-age` directive from one of the
+/// `Cache-Control` header fields, if present and syntactically valid.
+///
+/// The `Cache-Control` syntax permits multiple directives separated by commas
+/// (the canonical form) or, in some user agents, semicolons.  This helper
+/// therefore splits on both characters and returns the first `max-age` value
+/// it encounters that parses to a non-negative integer.  If multiple header
+/// fields are present the values are considered in header order.
+///
+/// The return value represents the freshness lifetime advertised by the
+/// response.  Callers needing more sophisticated freshness handling (Expires
+/// headers, heuristics, etc.) will need to layer their own logic on top of
+/// this primitive.
+///
+/// This logic was previously duplicated across several stateful caching rules;
+/// consolidating it here makes future maintenance easier.
+/// Where the member boundary is, and what counts as the directive being
+/// present, is this module's answer, one layer down; this function adds only
+/// the caching policy above it.
+pub fn get_cache_control_max_age(headers: &HeaderMap) -> Option<i64> {
+    // A section forbidding storage or unvalidated reuse advertises no freshness
+    // a caller could use, so a `max-age` written alongside is ignored.
+    if forbids_storage_or_reuse(headers) {
+        return None;
+    }
+    delta_seconds(headers, "max-age")
+}
+
+/// Extract the numeric value of an `s-maxage` directive from one of the
+/// `Cache-Control` header fields, if present and syntactically valid.
+///
+/// The semantics mirror `get_cache_control_max_age` but apply to shared caches
+/// only.  Private caches must ignore `s-maxage` when computing freshness
+/// lifetime, so this helper is mainly useful for rules verifying that clients
+/// or caches are not misapplying the directive.  As with the max-age helper,
+/// directives that explicitly forbid caching (`no-store`/`no-cache`) cause the
+/// value to be ignored.
+pub fn get_cache_control_s_maxage(headers: &HeaderMap) -> Option<i64> {
+    if forbids_storage_or_reuse(headers) {
+        return None;
+    }
+    delta_seconds(headers, "s-maxage")
+}
+
+/// Estimate the current age, in whole seconds, of a response observed at
+/// `observed_at` and being reasoned about at `now`.
+///
+/// This is a deliberate simplification of § 4.2.3's algorithm: it takes the
+/// `Age` the sender stated and adds the time the response has since spent in
+/// our own record, omitting the `response_delay` and `apparent_age` terms,
+/// which need request/response timing this linter does not record. The clamp of
+/// a negative elapsed to zero absorbs clock skew the way § 4.2.3 floors its own
+/// result. It is an estimate, and the two rules that ask are best-effort
+/// warnings about staleness — but they had each written it out, so they had two
+/// estimates, and only one of them said which.
+///
+/// A non-numeric or negative `Age` is dropped rather than propagated: the field
+/// is `delta-seconds`, so a value outside it states nothing about elapsed time.
+// cite(RFC 9111 § 5.1): "The "Age" response header field conveys the sender's estimate of the time since the response was generated or successfully validated at the origin server"
+// cite(RFC 9111 § 4.2.3): "corrected_initial_age = max(apparent_age, corrected_age_value)"
+pub fn estimated_age(
+    headers: &HeaderMap,
+    observed_at: chrono::DateTime<chrono::Utc>,
+    now: chrono::DateTime<chrono::Utc>,
+) -> i64 {
+    let stated_age = crate::helpers::headers::get_header_str(headers, "age")
+        .and_then(|s| s.trim().parse::<i64>().ok())
+        .filter(|seconds| *seconds >= 0)
+        .unwrap_or(0);
+    let resident = now.signed_duration_since(observed_at).num_seconds().max(0);
+    stated_age.saturating_add(resident)
+}
+
+/// Compute the freshness lifetime (in whole seconds) advertised by a response.
+/// This helper first consults `Cache-Control: max-age` (ignoring any values when
+/// `no-store` or `no-cache` are also present) and falls back to the `Expires`
+/// header if no max-age is available.  A missing or invalid value is treated as
+/// zero, representing an immediately stale entry.  Callers may use this to
+/// drive caching-related stateful rules without duplicating parsing logic.
+pub fn compute_freshness_lifetime(
+    headers: &HeaderMap,
+    resp_timestamp: chrono::DateTime<chrono::Utc>,
+) -> i64 {
+    // A section forbidding storage or unvalidated reuse is immediately stale
+    // whatever max-age or Expires says alongside it. The question is asked of
+    // the whole section, so splitting the directives across field lines cannot
+    // defeat it.
+    if forbids_storage_or_reuse(headers) {
+        return 0;
+    }
+
+    // The order is the sentence's order, and it is "use the first match". s-maxage is
+    // the arm above these two and is deliberately absent: it applies only to a shared
+    // cache, and this helper is not told whether it is one.
+    // cite(RFC 9111 § 4.2.1): "If the max-age response directive (Section 5.2.2.1) is present, use its value, or If the Expires response header field (Section 5.3) is present, use its value minus the value of the Date response header field (using the time the message was received if it is not present, as per Section 6.6.1 of [HTTP]), or"
+    if let Some(max_age) = get_cache_control_max_age(headers) {
+        return max_age;
+    }
+    if let Some(hv) = headers.get("expires") {
+        if let Ok(s) = hv.to_str() {
+            if let Ok(dt) = crate::http_date::parse_http_date_to_datetime(s.trim()) {
+                // Subtract Date, not the time we happened to receive this. The origin's
+                // clock is the one the sentence asks for, and `resp_timestamp` is only
+                // the fallback it names for when Date is absent.
+                // cite(RFC 9111 § 4.2.1): "Note that this calculation is intended to reduce clock skew by using the clock information provided by the origin server whenever possible."
+                let basis = headers
+                    .get("date")
+                    .and_then(|d| d.to_str().ok())
+                    .and_then(|d| crate::http_date::parse_http_date_to_datetime(d.trim()).ok())
+                    .unwrap_or(resp_timestamp);
+                let diff = dt.signed_duration_since(basis).num_seconds();
+                if diff > 0 {
+                    return diff;
+                }
+            }
+        }
+    }
+    0
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -356,5 +489,169 @@ mod tests {
     fn empty_members_are_not_directives() {
         let hm = headers(&[", ,max-age=1,"]);
         assert_eq!(directives(&hm).count(), 1);
+    }
+
+    #[test]
+    fn cache_control_max_age_helper_works() {
+        use super::get_cache_control_max_age;
+
+        let mut hm = HeaderMap::new();
+        assert!(get_cache_control_max_age(&hm).is_none());
+
+        hm.clear();
+        hm.append("cache-control", "max-age=10".parse().unwrap());
+        assert_eq!(get_cache_control_max_age(&hm), Some(10));
+
+        // semicolon separators should also parse
+        hm.clear();
+        hm.append("cache-control", "private; max-age=5".parse().unwrap());
+        assert_eq!(get_cache_control_max_age(&hm), Some(5));
+
+        // directive names are case-insensitive
+        hm.clear();
+        hm.append("cache-control", "Max-Age=20".parse().unwrap());
+        assert_eq!(get_cache_control_max_age(&hm), Some(20));
+        hm.clear();
+        hm.append("cache-control", "MAX-AGE=30".parse().unwrap());
+        assert_eq!(get_cache_control_max_age(&hm), Some(30));
+
+        // invalid values produce None, not previous result
+        hm.clear();
+        hm.append("cache-control", "max-age=xyz".parse().unwrap());
+        assert!(get_cache_control_max_age(&hm).is_none());
+
+        // negative values are ignored
+
+        // now exercises the new helper for s-maxage
+        use super::get_cache_control_s_maxage;
+
+        // no header
+        hm.clear();
+        assert!(get_cache_control_s_maxage(&hm).is_none());
+
+        hm.append("cache-control", "s-maxage=15".parse().unwrap());
+        assert_eq!(get_cache_control_s_maxage(&hm), Some(15));
+
+        // semicolon separators
+        hm.clear();
+        hm.append("cache-control", "public; s-maxage=7".parse().unwrap());
+        assert_eq!(get_cache_control_s_maxage(&hm), Some(7));
+
+        // case-insensitive name
+        hm.clear();
+        hm.append("cache-control", "S-MAXAGE=9".parse().unwrap());
+        assert_eq!(get_cache_control_s_maxage(&hm), Some(9));
+
+        // invalid value yields None
+        hm.clear();
+        hm.append("cache-control", "s-maxage=bad".parse().unwrap());
+        assert!(get_cache_control_s_maxage(&hm).is_none());
+
+        // negative is ignored
+        hm.clear();
+        hm.append("cache-control", "s-maxage=-1".parse().unwrap());
+        assert!(get_cache_control_s_maxage(&hm).is_none());
+
+        // no-store or no-cache cause helper to ignore
+        hm.clear();
+        hm.append("cache-control", "s-maxage=30, no-store".parse().unwrap());
+        assert!(get_cache_control_s_maxage(&hm).is_none());
+        hm.clear();
+        hm.append("cache-control", "no-cache, s-maxage=30".parse().unwrap());
+        assert!(get_cache_control_s_maxage(&hm).is_none());
+        hm.clear();
+        hm.append("cache-control", "max-age=-1".parse().unwrap());
+        assert!(get_cache_control_max_age(&hm).is_none());
+
+        // directives forbidding caching return None
+        let mut hm2 = HeaderMap::new();
+        hm2.append("cache-control", "max-age=30, no-store".parse().unwrap());
+        assert!(get_cache_control_max_age(&hm2).is_none());
+    }
+
+    #[test]
+    fn compute_freshness_lifetime_helper_works() {
+        use super::compute_freshness_lifetime;
+
+        let now = chrono::Utc::now();
+        let mut hm = HeaderMap::new();
+
+        // no freshness info -> zero
+        assert_eq!(compute_freshness_lifetime(&hm, now), 0);
+
+        // max-age takes precedence
+        hm.clear();
+        hm.append("cache-control", "max-age=30".parse().unwrap());
+        assert_eq!(compute_freshness_lifetime(&hm, now), 30);
+
+        // expires fallback
+        hm.clear();
+        let later = (now + chrono::Duration::seconds(20))
+            .format("%a, %d %b %Y %H:%M:%S GMT")
+            .to_string();
+        hm.append("expires", later.parse().unwrap());
+        let val = compute_freshness_lifetime(&hm, now);
+        assert!((val - 20).abs() <= 1, "got {} seconds", val);
+
+        // expires in past -> zero
+        hm.clear();
+        let past = (now - chrono::Duration::seconds(5))
+            .format("%a, %d %b %Y %H:%M:%S GMT")
+            .to_string();
+        hm.append("expires", past.parse().unwrap());
+        assert_eq!(compute_freshness_lifetime(&hm, now), 0);
+
+        // invalid date -> zero
+        hm.clear();
+        hm.append("expires", "not-a-date".parse().unwrap());
+        assert_eq!(compute_freshness_lifetime(&hm, now), 0);
+
+        // max-age still wins even if expires present
+        hm.clear();
+        hm.append("cache-control", "max-age=5".parse().unwrap());
+        hm.append("expires", later.parse().unwrap());
+        assert_eq!(compute_freshness_lifetime(&hm, now), 5);
+
+        // multiple cache-control headers: no-cache in one should zero out
+        hm.clear();
+        hm.append("cache-control", "max-age=100".parse().unwrap());
+        hm.append("cache-control", "no-cache".parse().unwrap());
+        assert_eq!(compute_freshness_lifetime(&hm, now), 0);
+
+        // Expires is measured from Date when Date is present, not from when we
+        // happened to receive the message. Here the origin's clock is 60s behind
+        // ours: Expires - Date is 20s, Expires - now is -40s. Reading it from `now`
+        // would call a fresh response stale.
+        hm.clear();
+        let origin_now = now - chrono::Duration::seconds(60);
+        let fmt =
+            |d: chrono::DateTime<chrono::Utc>| d.format("%a, %d %b %Y %H:%M:%S GMT").to_string();
+        hm.append("date", fmt(origin_now).parse().unwrap());
+        hm.append(
+            "expires",
+            fmt(origin_now + chrono::Duration::seconds(20))
+                .parse()
+                .unwrap(),
+        );
+        let val = compute_freshness_lifetime(&hm, now);
+        assert!(
+            (val - 20).abs() <= 1,
+            "expected ~20s from Date, got {}",
+            val
+        );
+
+        // With no Date, the time the message was received is the fallback the
+        // sentence names.
+        hm.clear();
+        hm.append(
+            "expires",
+            fmt(now + chrono::Duration::seconds(20)).parse().unwrap(),
+        );
+        let val = compute_freshness_lifetime(&hm, now);
+        assert!(
+            (val - 20).abs() <= 1,
+            "expected ~20s from receipt, got {}",
+            val
+        );
     }
 }
