+++ lint-http-rules/src/rules/immutable_cache_never_stale.rs
@@ -72,14 +72,14 @@ impl Rule for ImmutableCacheNeverStale {
 
             // Advertised freshness lifetime. The max-age/Expires calculation
             // (RFC 9111 §4.2.1) is owned by the helper, which carries the cite.
-            let freshness_lifetime = crate::helpers::headers::compute_freshness_lifetime(
+            let freshness_lifetime = crate::helpers::cache_control::compute_freshness_lifetime(
                 &prev_resp.headers,
                 prev_tx.timestamp,
             );
 
             // Age stated by the response plus the time it has since spent in our
             // record; the helper owns that estimate and what it leaves out.
-            let current_age = crate::helpers::headers::estimated_age(
+            let current_age = crate::helpers::cache_control::estimated_age(
                 &prev_resp.headers,
                 prev_tx.timestamp,
                 tx.timestamp,
